refactor #267 : 로그인 페이지 모바일/데스크탑 뷰 통합
